.ai-review/requirements.txt
anthropic>=0.40.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.0